Ejercicio 4 hecho sin redundancias

+++ BoletinExtra_2/src/main/java/org/example/Ej_4.java
@@ -7,11 +7,11 @@ public class Ej_4 {
         System.out.println("Cuantos dinero tienes?");
         dineroTotal = sc.nextInt();
         sc.close();
-        B100=(int)dineroTotal/100;
+        B100=dineroTotal/100;
         dineroTotal = dineroTotal-B100*100;
-        B20=(int)(dineroTotal/20);
+        B20=(dineroTotal/20);
         dineroTotal = dineroTotal-B20*20;
-        B5 = (int)dineroTotal/5;
+        B5 = dineroTotal/5;
         dineroTotal=dineroTotal-B5*5;
         M1=dineroTotal;
         System.out.println("Tienes "+B100+" billetes de 100€,"+B20+" billetes de 20€,"+B5+" billetes de 5€ y "+M1+" monedas de 1€");
